network_monitor/requirements.txt
pyyaml==6.0.3
matplotlib==3.10.8
seaborn==0.13.2
pandas==2.3.3
pytest==8.2.2
